+++ packages/trie/examples/merkle_patricia_trees/example1a.js
@@ -3,7 +3,7 @@
 const { Trie } = require('../../dist') // We import the library required to create a basic Merkle Patricia Tree
 
 const trie = new Trie() // We create an empty Merkle Patricia Tree
-console.log('Empty trie root (Bytes): ', trie.root) // The trie root (32 bytes)
+console.log('Empty trie root (Bytes): ', trie.root()) // The trie root (32 bytes)
 
 async function test() {
   const key = Buffer.from('testKey')
@@ -12,7 +12,7 @@ async function test() {
   const retrievedValue = await trie.get(key) // We retrieve (using "get") the value at key "testKey"
   console.log('Value (Bytes): ', retrievedValue)
   console.log('Value (String): ', retrievedValue.toString())
-  console.log('Updated trie root:', trie.root) // The new trie root (32 bytes)
+  console.log('Updated trie root:', trie.root()) // The new trie root (32 bytes)
 }
 
 test()